Refactoring - call register_object in init() method Move commands out of Config and AliasManager so they do not have to subclass plugin Add AliasManager to Plugin context so that it can be inspected PEP-8 fixes
