+++ abacura/plugins/aliases/manager.py
@@ -4,19 +4,21 @@ import csv
 import io
 import re
 from typing import Dict, TYPE_CHECKING
-from rich.panel import Panel
-from rich.pretty import Pretty
-from abacura.plugins import Plugin, command
+
+from serum import inject
 
 if TYPE_CHECKING:
     from abacura import Session
     from abacura.config import Config
 
+
+@inject
 class AliasManager:
     """Alias manager"""
 
     session: Session
     config: Config
+
     def __init__(self):
         super().__init__()
         self.aliases: Dict[str, str] = {}
@@ -34,7 +36,7 @@ class AliasManager:
 
                     file_like = io.StringIO(alias_line)
                     parts = next(csv.reader(file_like, delimiter=' '))
-                    #self.session.output(f"[bold yellow] PREPARSE: {list(parts)}", markup = True)
+                    # self.session.output(f"[bold yellow] PREPARSE: {list(parts)}", markup = True)
                     parsed = []
                     for token in parts:
                         m = self.re_param.match(token)
@@ -42,11 +44,11 @@ class AliasManager:
                             parsed.append(args[int(m.group(1))] if int(m.group(1)) < len(args) else r'')
                         else:
                             parsed.append(token)
-                    #self.session.output(f"Parsed: {parsed}")
+                    # self.session.output(f"Parsed: {parsed}")
                     parsed_alias = ' '.join(parsed)
 
                     # SEND
-                    #self.session.output(f"[bold yellow] SEND: {parsed_alias}", markup = True)
+                    # self.session.output(f"[bold yellow] SEND: {parsed_alias}", markup = True)
                     self.session.player_input(parsed_alias)
 
             except StopIteration:
@@ -54,13 +56,3 @@ class AliasManager:
 
             return True
         return False
-    
-    @command(name="alias")
-    def alias(self):
-        """list, remove add aliases"""
-        buf = "[bold white]Aliases:\n"
-        for key in self.aliases.items():
-            self.session.output(Pretty(key), actionable=False)
-            buf += f"{key[0]}: {key[1]}\n"
-        
-        self.session.output(Panel(buf), actionable=False)
